add sensor readings to web

+++ hemcs/datahandler.cpp
@@ -1,7 +1,7 @@
 #include "datahandler.h"
 
 
-StaticJsonDocument<200> graphSensorReading(double sensor_data[], size_t data_size) {
+StaticJsonDocument<200> graphSensorReading(float sensor_data[], size_t data_size) {
     const uint8_t cmd = 2;
     StaticJsonDocument<200> root;
     root["cmd"] = cmd;
