+++ hemcs/datahandler.h
@@ -3,6 +3,6 @@
 
 #include <ArduinoJson.h>
 
-StaticJsonDocument<200> graphSensorReading(double sensor_data[], size_t data_size);
+StaticJsonDocument<200> graphSensorReading(float sensor_data[], size_t data_size);
 
 #endif //DATAHANDLER_H
